Made readable and writable stores separate
