+++ src/main/java/io/github/csgroup/quizmaker/utils/stores/StoreListener.java
@@ -1,6 +1,8 @@
 package io.github.csgroup.quizmaker.utils.stores;
 
+import io.github.csgroup.quizmaker.utils.stores.readable.ReadableStore;
+
 public interface StoreListener<T> 
 {
-	public void onStoreChange(Store<T> store, T value);
+	public void onStoreChange(ReadableStore<T> store, T value);
 }
