+++ src/main/java/io/github/csgroup/quizmaker/utils/stores/readable/AbstractReadableStore.java
@@ -1,19 +1,12 @@
-package io.github.csgroup.quizmaker.utils.stores;
+package io.github.csgroup.quizmaker.utils.stores.readable;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public abstract class AbstractStore<T> implements Store<T>
+import io.github.csgroup.quizmaker.utils.stores.StoreListener;
+
+public abstract class AbstractReadableStore<T> implements ReadableStore<T>
 {
-	@Override
-	public void set(T value)
-	{
-		setValue(value);
-		
-		fireChange(value);
-	}
-	
-	protected abstract void setValue(T value);
 	
 	@Override
 	public abstract T get();
