+++ src/main/java/io/github/csgroup/quizmaker/utils/stores/writable/AbstractWritableStore.java
@@ -1,0 +1,16 @@
+package io.github.csgroup.quizmaker.utils.stores.writable;
+
+import io.github.csgroup.quizmaker.utils.stores.readable.AbstractReadableStore;
+
+public abstract class AbstractWritableStore<T> extends AbstractReadableStore<T> implements WritableStore<T>
+{
+	@Override
+	public void set(T value)
+	{
+		setValue(value);
+		
+		fireChange(value);
+	}
+	
+	protected abstract void setValue(T value);
+}
