+++ src/main/java/io/github/csgroup/quizmaker/utils/stores/writable/WritableStore.java
@@ -1,0 +1,8 @@
+package io.github.csgroup.quizmaker.utils.stores.writable;
+
+import io.github.csgroup.quizmaker.utils.stores.readable.ReadableStore;
+
+public interface WritableStore<T> extends ReadableStore<T> 
+{
+	public void set(T value);
+}
